pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ebook-downloader"
version = "0.1.0"
description = "城通网盘电子书批量下载工具"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "aiosqlite>=0.20.0",
]

[project.scripts]
ebook-downloader = "ebook_downloader.cli:main"
